+++ src/controls/folderExplorer/NewFolder/INewFolderState.ts
@@ -1,0 +1,5 @@
+export interface INewFolderState {
+  folderName: string;
+  showInput: boolean;
+  loading: boolean;
+}