+++ src/index.ts
@@ -11,6 +11,8 @@ export * from './ChartControl';
 export * from './Progress';
 export * from './DateTimePicker';
 export * from './FilePicker';
+export * from './FolderExplorer';
+export * from './IconPicker';
 
 export * from './IFrameDialog';
 export * from './IFramePanel';